+++ src/components/displays/Skills.js
@@ -10,7 +10,6 @@ const Skills = () => {
   return (
     <SkillsContainer>
       {icons.skills.map((icon, idx) => {
-        console.log(idx, icon)
         return <SkillCard key={idx} icon={icon} />;
       })}
     </SkillsContainer>
